CI: Fix running large test suite in merge requests
Using a relative path for the large test suite's test files only works
in CI if the merge request comes from a branch in the upstream
repository. For merge requests from other accounts/orgs, it won't find
the test files unless that is also forked.

This commit swaps over to an absolute HTTPS URL for the testfiles repo.
The SSH (git@) kind also wouldn't work because authentication fails.